move AppointmentComparator
